fix(analyses): correct statistical inefficiency estimator
estimate_correlation_time(method="integration") trapezoid-integrated the
normalised ACF from lag zero, so the lag-zero term was counted a second
time when g = 1 + 2*tau/dt was applied, and it floored tau at one
timestep. White noise therefore returned g = 3.00 instead of 1.00 and
correlated series came back about +2 too high. Every sem_rmsd, sem_rg,
sem_sasa, sem_distance and n_independent_frames written to an artifact
inherited that bias, and the correct pymbar-style statistical_inefficiency
in the same module was called by nothing.

The integration path now sums g = 1 + 2*sum C(t)*(1 - t/N) over positive
lags with the first-non-positive cutoff and derives tau = (g - 1)/2 * dt,
sharing one helper with statistical_inefficiency so the two can no longer
disagree. The floor is gone, so g may approach 1. n_independent is now the
float N/g rather than an int-truncated count, and the five plugin payloads
that carry it are typed float. The first_zero and exponential_fit methods
are withdrawn; on an AR(1) series with phi = 0.9 they were off by factors
of five and two, and they raise ValueError now. The module docstring drops
the Flyvbjerg and Petersen block-averaging citation, since no block
averaging exists here, notes that the N_independent < 10 threshold is a
house convention rather than a LiveCoMS rule, and gains a full References
section.

White noise with N = 20000 now gives g = 1.04 where it gave 3.00. AR(1)
with phi = 0.5 gives 3.02 against an analytic 3.00, and phi = 0.9 gives
19.8 against an analytic 19.0.

+++ src/polyzymd/analyses/catalytic_triad/_mda.py
@@ -605,7 +605,7 @@ def _simultaneous_contact_statistics(
             "n_independent": None,
             "warning": str(exc),
         }
-    n_independent = int(tau_result.n_independent)
+    n_independent = float(tau_result.n_independent)
     sem = (
         float(np.sqrt(fraction * (1.0 - fraction) / float(n_independent)))
         if n_independent > 0
